requirements.txt
annotated-types==0.7.0
anyio==4.10.0
azure-functions==1.23.0
certifi==2025.8.3
charset-normalizer==3.4.3
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.27
langchain-core==0.3.76
langchain-text-splitters==0.3.11
langsmith==0.4.28
logging==0.4.9.6
MarkupSafe==3.0.2
orjson==3.11.3
packaging==25.0
pydantic==2.11.9
pydantic_core==2.33.2
PyYAML==6.0.2
requests==2.32.5
requests-toolbelt==1.0.0
sniffio==1.3.1
SQLAlchemy==2.0.43
tenacity==9.1.2
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
Werkzeug==3.1.3
zstandard==0.25.0
fastapi
faiss-cpu
langchain_openai